requirements.txt
Django
Pillow
djangorestframework
markdown
django-filter
Unidecode
unicodedata2
django-taggit
django-taggit-serializer
django-currentuser
django-crispy-forms
sorl-thumbnail
django-recaptcha
django-cookie-law
django-admin-interface
pandas
django-hitcount
django-summernote